Fix HTTP::PUT request and uri port.  Need to use port 88 to sneak around varnish which is objecting to put.

+++ common/src/network/uri.hpp
@@ -127,7 +127,7 @@ class Uri {
       ss << hostname_;
     }
     if (port_) {
-      ss << ":" << port_;
+      ss << ":" << (uint32_t)port_;
     }
     ss << path_;
     if (queryParameters_.size()) {
